Added find collision func to registry
Added find collision func to registry now it just uses printf to output
where the collisions are. it seems to work with 100% accuracy.
Collisions are output twice since it checks the position for each
character there. as such if it appears 3 time there'll be letters there.
We'll probably want to make multiple sweeps eventually